Rename auto http interception modules & artifacts to contain word auto
