+++ agent/build.gradle.kts
@@ -24,8 +24,8 @@ dependencies {
     // TODO api(project(":integration:session-recording"))
     api(project(":integration:crash"))
     api(project(":integration:anr"))
-    api(project(":integration:httpurlconnection"))
-    api(project(":integration:okhttp3"))
+    api(project(":integration:httpurlconnection-auto"))
+    api(project(":integration:okhttp3-auto"))
     api(project(":integration:startup"))
     api(project(":integration:interactions"))
     api(project(":integration:lifecycle"))
