+++ app/build.gradle.kts
@@ -6,8 +6,8 @@ plugins {
     id("com.android.application")
     id("kotlin-android")
     // Uncomment this to test HttpURLConnection instrumentation
-    //id("com.splunk.android.rum-okhttp3-plugin") version "24.4.1"
-    //id("com.splunk.android.rum-httpurlconnection-plugin") version "24.4.1"
+    //id("com.splunk.android.rum-okhttp3-auto-plugin") version "24.4.1"
+    //id("com.splunk.android.rum-httpurlconnection-auto-plugin") version "24.4.1"
 }
 
 apply<ConfigAndroidApp>()
