+++ instrumentation/buildtime/httpurlconnection-auto/plugin/src/main/assets/LICENSE.txt
