+++ integration/httpurlconnection-auto/lint-baseline.xml
