+++ integration/httpurlconnection-auto/src/main/assets/LICENSE.txt
