--- integration/httpurlconnection/src/main/AndroidManifest.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <provider
-            android:name="com.splunk.rum.integration.httpurlconnection.HttpURLInstaller"
-            android:authorities="${applicationId}.HttpURLInstaller"
-            android:enabled="true"
-            android:exported="false" />
-
-    </application>
-
-</manifest>
