--- integration/httpurlconnection/src/main/java/com/splunk/rum/integration/httpurlconnection/HttpURLAdditionalAttributesExtractor.kt
@@ -1,59 +1,0 @@
-/*
- * Copyright 2025 Splunk Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.splunk.rum.integration.httpurlconnection
-
-import com.splunk.sdk.common.otel.internal.RumConstants
-import com.splunk.sdk.utils.ServerTimingHeaderParser
-import io.opentelemetry.api.common.AttributesBuilder
-import io.opentelemetry.context.Context
-import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor
-import java.net.URLConnection
-
-class HttpURLAdditionalAttributesExtractor: AttributesExtractor<URLConnection, Int> {
-
-    override fun onStart(attributes: AttributesBuilder, parentContext: Context, connection: URLConnection) {
-        attributes.put(RumConstants.COMPONENT_KEY, RumConstants.COMPONENT_HTTP)
-    }
-
-    override fun onEnd(
-        attributes: AttributesBuilder,
-        context: Context,
-        connection: URLConnection,
-        responseCode: Int?,
-        error: Throwable?
-    ) {
-        onResponse(attributes, connection)
-    }
-
-    private fun onResponse(attributes: AttributesBuilder, connection: URLConnection) {
-        connection.headerFields.forEach { header ->
-            if (!header.key.equals(RumConstants.SERVER_TIMING_HEADER, ignoreCase = true)) {
-                return@forEach
-            }
-
-            // HttpURLConnection consolidates multiple headers with the same name into a single header entry,
-            // combining their values into a list. We want to capture the last valid server-timing header.
-            header.value.forEach { headerValue ->
-                val serverTraceContext = ServerTimingHeaderParser.parse(headerValue)
-                serverTraceContext?.let {
-                    attributes.put(RumConstants.LINK_TRACE_ID_KEY, it.traceId)
-                    attributes.put(RumConstants.LINK_SPAN_ID_KEY, it.spanId)
-                }
-            }
-        }
-    }
-}
